Issue 112: updated to latest version of hosting.com
git-svn-id: http://jclouds.googlecode.com/svn/trunk@2372 3d8758e0-26b5-11de-8745-db77d3ebf521

+++ vcloud/terremark/src/test/java/org/jclouds/vcloud/terremark/TerremarkVCloudClientLiveTest.java
@@ -203,8 +203,8 @@ public class TerremarkVCloudClientLiveTest extends VCloudClientLiveTest {
                memory);
       assertEquals(vApp.getResourceAllocationByType().get(ResourceType.DISK_DRIVE)
                .getVirtualQuantity(), hardDisk);
-      assertEquals(vApp.getSize(), vApp.getResourceAllocationByType().get(ResourceType.DISK_DRIVE)
-               .getVirtualQuantity());
+      assertEquals(vApp.getSize().longValue(), vApp.getResourceAllocationByType().get(
+               ResourceType.DISK_DRIVE).getVirtualQuantity());
    }
 
    private void doCheckPass(InetAddress address) throws IOException {
