+++ vcloud/terremark/src/test/java/org/jclouds/vcloud/terremark/TerremarkVCloudComputeClientLiveTest.java
@@ -149,7 +149,7 @@ public class TerremarkVCloudComputeClientLiveTest {
                memory);
       assertEquals(vApp.getResourceAllocationByType().get(ResourceType.DISK_DRIVE)
                .getVirtualQuantity(), hardDisk);
-      assertEquals(vApp.getSize(), vApp.getResourceAllocationByType().get(ResourceType.DISK_DRIVE)
+      assertEquals(vApp.getSize().longValue(), vApp.getResourceAllocationByType().get(ResourceType.DISK_DRIVE)
                .getVirtualQuantity());
    }
